requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
sqlmodel
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
pydantic[email]
python-multipart